Query VMs' hard drives info from libvirt

+++ rebuild-vm.sh
@@ -1,7 +1,6 @@
 #!/bin/sh
 set -e
 MYSELF="${0##*/}"
-VM_OS_VG=as-ubuntu-vg
 CALLBACK_PORT=3333
 
 vm="$1"
@@ -18,20 +17,27 @@ if [ ! -f "$UBUNTU_IMG" ]; then
 	wget -N -O "$UBUNTU_IMG" "$UBUNTU_IMG_URL"
 fi
 
-case $vm in
-	saceph-adm|saceph-rgw)
-		VM_OS_VG="wdgreen"
-		;;
-	*)
-		;;
-esac
-
 cd "${0%/*}"
 virsh destroy "$vm" || true
 if ! virsh domid "$vm" >/dev/null 2>&1; then
 	virsh define "${vm}.xml"
 fi
-vm_hdd="/dev/$VM_OS_VG/${vm}-os"
+
+vm_drives="`./get-vm-harddrives $vm`"
+vm_hdd=''
+for hd in $vm_drives; do
+	case $hd in
+		*-os)
+			vm_hdd="$hd"
+			;;
+	esac
+done
+
+if [ -z "$vm_hdd" ]; then
+	echo "${MYSELF}: could not find VM hdd backing device" >&2
+	exit 1
+fi
+
 sudo chgrp adm "$vm_hdd"
 dd if=/dev/zero of="$vm_hdd" bs=1M count=32
 virt-resize --expand /dev/sda1 "$UBUNTU_IMG" "$vm_hdd"
